Added a new implementation of the fish population.
The implementation largely focusses on representing
the number of fish as an array of fish count where
each index represents the days left and the value for
that index represents the number of fish that have this
particular amount of days left.

The implementation is much more efficient; due to the
exponential nature of the problem, the implementation
that was done in part 1 of the day cannot be used for
part 2 since it would run hours or even days due to
the exponential growth.